feat: rebuild frontend with new component architecture and v--v brand
Replace diary/food-library structure with focused day-planning UI: new components
(AIPanel, Cart, FoodModal, HeroSearch, ResultsGrid, Nav, Toast, FilterBar, FoodGlyph),
custom hooks (useToast, useKeyboardShortcut), brand styles, and consolidated types.
Remove old hooks and pages that no longer match the app design.

+++ server/src/db.ts
@@ -18,10 +18,10 @@ export const initDb = () => {
     CREATE TABLE IF NOT EXISTS foods (
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       name TEXT NOT NULL UNIQUE,
-      calories_per_100g INTEGER NOT NULL,
-      protein_per_100g INTEGER NOT NULL,
-      carbs_per_100g INTEGER NOT NULL,
-      fat_per_100g INTEGER NOT NULL,
+      calories_per_100g REAL NOT NULL,
+      protein_per_100g REAL NOT NULL,
+      carbs_per_100g REAL NOT NULL,
+      fat_per_100g REAL NOT NULL,
       created_at TEXT NOT NULL DEFAULT (datetime('now'))
     );
